+++ core/src/main/java/io/cucumber/stepexpression/TypeResolver.java
@@ -8,7 +8,9 @@ import java.lang.reflect.Type;
 public interface TypeResolver {
 
     /**
-     * A type to data convert the table or doc string to. May not return null.
+     * A type to data convert the table or doc string to.
+     *
+     * May return null if no type could be resolved.
      *
      * @return a type
      */